Fix builds with `rust-toolchain.toml`
When a `rust-toolchain.toml`-file is present in the project directory,
any calls to rust tools through `rustup` from inside the project
correctly pick up the set toolchain version.
However, when `sccache rustup` is called by `rustc_version`,
`sccache` executes `rustc` in the system root,
where no `rust-toolchain.toml` is present.
That would previously lead to `rustup` defaulting to `stable`,
which was removed from the docker container.
This fix sets the `RUSTUP_TOOLCHAIN` environment variable
to globally set the toolchain to the defined version.
This also leads to the correct rust-version being reported
in both the docker and host builds.

Fixes #348